+++ app/schemas/credit_card_schemas.py
@@ -4,12 +4,12 @@ from pydantic import BaseModel
 
 class NewCreditCardReq(BaseModel):
     name: str
-    limit: float
+    limit: int
     main_credit_card_id: int | None
 
 class CreditCardReq(BaseModel):
     name: str | None = None
-    limit: float | None = None
+    limit: int | None = None
     main_credit_card_id: int | None = None
     user_id: int
 
@@ -17,7 +17,7 @@ class CreditCardReq(BaseModel):
 class CreditCardRes(BaseModel):
     id: int
     name: str
-    limit: float
+    limit: int
     main_credit_card_id: int | None
     user_id: int
     created_at: datetime
